requirements.txt
"""
aiogram==3.0.0
python-dotenv==1.0.0
sqlalchemy==2.0.0
asyncpg==0.27.0
requests==2.31.0
openai==1.2.3
"""
